refactor Stack more to my taste
I eliminated OStack and UStack, in favor of StackName and a new
typedef UStackName, and a Deck newtype to handle the weird deck
semantics.  IStack is still in place and the gain situation is
unchanged.

I left most of the classes in place, but removed the InputStack class,
replacing it with one a set of hooks that get called when cards are
added to stacks.  This makes it runtime-modifiable (which I *think* is
a nice feature), and also removes our one remaining "unsafe"
stack-modifying function.  We could definitely benefit from a nicer
"announcement" framework.

Tested-on: morland i686 GNU/Linux